+++ spec/factories/categories.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+FactoryBot.define do
+  factory :category do
+    sequence(:name) { |n| "category #{n}" }
+  end
+end